backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
openai
langsmith
pydantic
pydantic-settings
httpx
python-multipart
pypdf
google-api-python-client
google-auth-oauthlib
apscheduler
